WalletExplorer code cleanup
Moved classes to own file.
Fixed spacing.
Using object initializer in WalletViewModel ctor
