+++ WalletWasabi.Gui/Controls/WalletExplorer/CoinJoinActionViewModel.cs
@@ -1,0 +1,12 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace WalletWasabi.Gui.Controls.WalletExplorer
+{
+	public class CoinJoinActionViewModel : WalletActionViewModel
+	{
+		public CoinJoinActionViewModel(WalletViewModel walletViewModel)
+			: base("CoinJoin", walletViewModel) { }
+	}
+}
