+++ WalletWasabi.Gui/Controls/WalletExplorer/ReceiveActionViewModel.cs
@@ -1,0 +1,12 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace WalletWasabi.Gui.Controls.WalletExplorer
+{
+	public class ReceiveActionViewModel : WalletActionViewModel
+	{
+		public ReceiveActionViewModel(WalletViewModel walletViewModel)
+			: base("Receive", walletViewModel) { }
+	}
+}
